fixing windows build issues.
git-svn-id: http://svn.code.sf.net/p/calchart/svn/trunk@336 eb3ee358-668b-4427-8ffb-f370b80720d0

+++ src/cont.cpp
@@ -1139,7 +1139,7 @@ void ContProcRotate::Compile(AnimateCompile* anim)
 	anim->Append(boost::shared_ptr<AnimateCommand>(new AnimateCommandRotate((unsigned)ABS(b), c,
 // Don't use Magnitude() because
 // we want Coord numbers
-		sqrt(rad.x*rad.x + rad.y*rad.y),
+		sqrt(static_cast<float>(rad.x*rad.x + rad.y*rad.y)),
 		start_ang, start_ang+angle,
 		backwards)),
 		this);
